+++ src/pynescript/__about__.py
@@ -30,4 +30,4 @@ package identity.
 from __future__ import annotations
 
 
-__version__ = "0.6.2"
+__version__ = "0.6.3"